gathering balance, deposit data type change

See merge request s12-fintech-finance-sub1/S12P21B110!118

+++ server/src/main/java/com/b110/jjeonchongmu/domain/gathering/dto/MemberManageResponseDTO.java
@@ -36,7 +36,7 @@ public class MemberManageResponseDTO {
         private String name;
         private String email;
         private LocalDateTime createdAt;
-        private Integer balance;
+        private Long balance;
         private Boolean gatheringPaymentStatus;
     }
 
@@ -48,7 +48,7 @@ public class MemberManageResponseDTO {
         private String name;
         private String email;
         private LocalDateTime createdAt;
-        private Integer balance;
+        private Long balance;
         private Boolean gatheringPaymentStatus;
     }
 }
